Move rubriken - schlagwörter tests to separate file
